requirements.txt
Flask==2.3.3
python-nmap==0.7.1
scapy==2.5.0
